+++ master-nafsdm/webinterface/logPath.py
@@ -1,0 +1,1 @@
+logPath = "../webinterface_log.log"