+++ setupMaster.sh
@@ -10,18 +10,21 @@ fi
 
 #  DL_VERSION will be changed at the time of update
 DL_URL="https://github.com/MrKaKisen/nafsdm/archive/"
+REQ_URL="https://raw.githubusercontent.com/MrKaKisen/nafsdm/master/scripts/requirements_master.txt"
 GITHUB_DIR="master-nafsdm"
 HOME_DIR="/home/master-nafsdm"
 USER="master-nafsdm"
 
 echo "###################################################################"
-echo "THIS SCRIPT WILL NOT WORK FOR UPDATING YOUR INSTALLATION"
-echo "Welcome to nafsdm master install! Please enter your operating system name ('debian', 'ubuntu' and 'centos' only supported)"
-echo -n "Operating system: "
+echo "* nafsdm-master installation script"
+echo "* note: this installer will not upgrade your installation"
+echo "###################################################################"
+echo "* Please enter your operating system name ('debian', 'ubuntu' and 'centos' only supported)"
+echo -n "* Operating system: "
 read OPERATINGSYS
 
 if [ "$OPERATINGSYS" == "centos" ]; then
-  echo "Installing packages.."
+  echo "* Installing packages.."
   yum update -y
   yum install python curl wget -y
 
@@ -30,49 +33,57 @@ if [ "$OPERATINGSYS" == "centos" ]; then
   python get-pip.py
   rm get-pip.py -rf
 
-  pip install requests
+  cd /tmp
+  wget -O requirements.txt $REQ_URL
+
+  pip install -r requirements.txt
+  rm -rf requirements.txt
 elif [[ "$OPERATINGSYS" == "debian" ]] || [[ "$OPERATINGSYS" == "ubuntu" ]] ; then
-  echo "Installing packages.."
+  echo "* Installing packages.."
   apt-get update -y
   apt-get install python python-pip curl wget -y
 
-  pip install requests
+  cd /tmp
+  wget -O requirements.txt $REQ_URL
+
+  pip install -r requirements.txt
+  rm -rf requirements.txt
 else
   echo "Invalid operating system. Only 'debian', 'ubuntu' and 'centos' supported."
   exit 1
 fi
 
 # get which version is the latest
-echo "Fetching information about latest version.."
+echo "* Fetching information about latest version.."
 LATEST_VERSION=$(curl https://raw.githubusercontent.com/MrKaKisen/nafsdm/master/version.txt)
 
 # select version
-echo "Please select your version. Type in the version number or type 'latest' for latest version."
-echo -n "Version: "
+echo "* Please select your version. Type in the version number or type 'latest' for latest version."
+echo -n "* Version: "
 read VERSION_USER
 
 if [ "$VERSION_USER" == "latest" ]; then
-  echo -n "Confirm? (y/n): "
+  echo -n "* Confirm? (y/n): "
   read CONFIRM
   if [ "$CONFIRM" == "y" ]; then
     DL_VERSION="$LATEST_VERSION"
   else
-    echo "Aborting.."
+    echo "* Aborting.."
     exit 1
   fi
 else
-  echo -n "Confirm? If version doesn't exist, script will fail. (y/n): "
+  echo -n "* Confirm? If version doesn't exist, script will fail. (y/n): "
   read CONFIRM
   if [ "$CONFIRM" == "y" ]; then
     DL_VERSION="$VERSION_USER"
   else
-    echo "Aborting.."
+    echo "* Aborting.."
     exit 1
   fi
 fi
 
-echo "Required packages installed!"
-echo "Downloading nafsdm & installing.."
+echo "* Required packages installed!"
+echo "* Downloading nafsdm & installing.."
 
 # download in temp dir
 cd /tmp
@@ -93,11 +104,19 @@ cp /tmp/nafsdm/LICENSE $HOME_DIR/LICENSE
 cp /tmp/nafsdm/systemconfigs/nafsdmctl /usr/bin/nafsdmctl
 cp /tmp/nafsdm/systemconfigs/nafsdm-master /usr/bin/nafsdm-master
 
+# nafsdm webinterface
+cp /tmp/nafsdm/systemconfigs/nafsdm-webinterface.service /home/master-nafsdm/webinterface/nafsdm-webinterface.service
+chmod +x /home/master-nafsdm/webinterface/enableInterface.sh
+chmod +x /home/master-nafsdm/webinterface/start.sh
+
 chmod +x /usr/bin/nafsdmctl
 chmod +x /usr/bin/nafsdm-master
 
-echo "Installed. Cleanup.."
+echo "* Installed. Cleanup.."
 
 rm /tmp/nafsdm -rf
-
-echo "To continue, please run 'nafsdm-master' for first time setup."
+echo "###################################################################"
+echo "* nafsdm-master installation complete"
+echo "* To continue, please run 'nafsdm-master' for first time setup."
+echo "* Thank you for using nafsdm!"
+echo "###################################################################"